requirements.txt
bonobo
requests
redis
psycopg2
python-dotenv
tornado
python-socketio[client]